console: Refactor connect in fw version select
