+++ pkg/webui/console/containers/device-profile-section/device-selection/fw-version-select/index.js
@@ -12,9 +12,69 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import connect from './connect'
-import FirmwareVersionSelect from './fw-version-select'
+import React from 'react'
+import { defineMessages } from 'react-intl'
+import { useSelector } from 'react-redux'
+import { useParams } from 'react-router-dom'
 
-const ConnectedFirmwareVersionSelect = connect(FirmwareVersionSelect)
+import Field from '@ttn-lw/components/form/field'
+import Select from '@ttn-lw/components/select'
+import { useFormContext } from '@ttn-lw/components/form'
 
-export { ConnectedFirmwareVersionSelect as default, FirmwareVersionSelect }
+import PropTypes from '@ttn-lw/lib/prop-types'
+import sharedMessages from '@ttn-lw/lib/shared-messages'
+
+import { isUnknownHwVersion, SELECT_OTHER_OPTION } from '@console/lib/device-utils'
+
+import { selectDeviceModelFirmwareVersions } from '@console/store/selectors/device-repository'
+
+const m = defineMessages({
+  title: 'Firmware Ver.',
+})
+
+const formatOptions = (versions = []) =>
+  versions
+    .map(version => ({
+      value: version.version,
+      label: version.version,
+    }))
+    .concat([{ value: SELECT_OTHER_OPTION, label: sharedMessages.otherOption }])
+
+const FirmwareVersionSelect = props => {
+  const { brandId, modelId, hwVersion } = useParams()
+  const { name, onChange, ...rest } = props
+  const { setFieldValue } = useFormContext()
+
+  const versions = useSelector(state =>
+    selectDeviceModelFirmwareVersions(state, brandId, modelId).filter(
+      ({ supported_hardware_versions = [] }) =>
+        (Boolean(hwVersion) && supported_hardware_versions.includes(hwVersion)) ||
+        // Include firmware versions when there are no hardware versions configured in device repository
+        // for selected end device model.
+        isUnknownHwVersion(hwVersion),
+    ),
+  )
+
+  const options = React.useMemo(() => formatOptions(versions), [versions])
+
+  React.useEffect(() => {
+    if (options.length > 0 && options.length <= 2) {
+      setFieldValue('version_ids.firmware_version', options[0].value)
+    }
+  }, [setFieldValue, options])
+
+  return (
+    <Field {...rest} options={options} name={name} title={m.title} component={Select} autoFocus />
+  )
+}
+
+FirmwareVersionSelect.propTypes = {
+  name: PropTypes.string.isRequired,
+  onChange: PropTypes.func,
+}
+
+FirmwareVersionSelect.defaultProps = {
+  onChange: () => null,
+}
+
+export default FirmwareVersionSelect
