+++ pkg/webui/locales/en.json
@@ -501,6 +501,7 @@
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.title": "End device brand",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.noOptionsMessage": "No matching brand found",
   "console.containers.device-profile-section.device-selection.fw-version-select.fw-version-select.title": "Firmware Ver.",
+  "console.containers.device-profile-section.device-selection.fw-version-select.index.title": "Firmware Ver.",
   "console.containers.device-profile-section.device-selection.hw-version-select.hw-version-select.title": "Hardware Ver.",
   "console.containers.device-profile-section.device-selection.model-select.model-select.noOptionsMessage": "No matching model found",
   "console.containers.device-profile-section.hints.other-hint.hintTitle": "Your end device will be added soon!",
